gitignore, move zlib file, update filename to versionhelpers
