+++ src/inc/internal/VersionHelpers.hpp
