+++ src/msix/CMakeLists.txt
@@ -134,7 +134,7 @@ if(MSIX_PACK)
         pack/BundleManifestWriter.cpp
         pack/BundleWriterHelper.cpp
         pack/AppxBundleWriter.cpp
-        pack/GeneralUtil.cpp
+        pack/VersionHelpers.cpp
     )
 endif()
